+++ viscid/vutil.py
@@ -757,7 +757,7 @@ def convert_deprecated_floats(value, varname="value"):
         value = "{0}f".format(value)
     return value
 
-def prepare_lines(lines, scalars=None, do_connections=False):
+def prepare_lines(lines, scalars=None, do_connections=False, other=None):
     """Concatenate and standardize a list of lines
 
     Args:
@@ -769,11 +769,14 @@ def prepare_lines(lines, scalars=None, do_connections=False):
             or 3xN for an rgb color for each point. If the shape is
             1xNlines, the scalar is broadcast so the whole line gets
             the same value, and likewise for 3xNlines and rgb colors.
+            Can also be a list of hex color (#ffffff) strings.
             Otherwise, scalars is reshaped to -1xN.
         do_connections (bool): Whether or not to make connections array
+        other (dict): a dictionary of other arrays that should be
+            reshaped and the like the same way scalars is
 
     Returns:
-        (lines, scalars, connections)
+        (lines, scalars, connections, dict)
 
         * lines (ndarray): 3xN array of N xyz points. N is the sum of
             the lengths of all the lines
@@ -781,6 +784,7 @@ def prepare_lines(lines, scalars=None, do_connections=False):
             rgb values, or None
         * connections (ndarray): 2xN array describing the forward and
             backward connectedness of the lines, or None
+        * other (dict): a dict of N length arrays
 
     Raises:
         ValueError: If rgb data is not in a valid range or the shape
@@ -788,6 +792,7 @@ def prepare_lines(lines, scalars=None, do_connections=False):
     """
     nlines = len(lines)
     npts = [line.shape[1] for line in lines]
+    N = np.sum(npts)
     first_idx = np.cumsum([0] + npts[:-1])
     lines = [np.asarray(line) for line in lines]
     lines = np.concatenate(lines, axis=1)
@@ -797,7 +802,7 @@ def prepare_lines(lines, scalars=None, do_connections=False):
             viscid.logger.warn("Overriding line scalars with scalars kwarg")
         else:
             scalars = lines[3:, :]
-            lines = lines[:3, :]
+        lines = lines[:3, :]
 
     if scalars is not None:
         scalars = np.atleast_2d(scalars)
@@ -805,7 +810,11 @@ def prepare_lines(lines, scalars=None, do_connections=False):
             # one scalar for each line, so broadcast it
             scalars = scalars.reshape(nlines, 1)
             scalars = [scalars[i].repeat(ni) for i, ni in enumerate(npts)]
-            scalars = np.concatenate(scalars, axis=0).reshape(1, np.sum(npts))
+            scalars = np.concatenate(scalars, axis=0).reshape(1, N)
+        elif scalars.shape == (N, 1) or scalars.shape == (1, N):
+            # catch these so they're not interpreted as colors if
+            # nlines == 1 and N == 3; ie. 1 line with 3 points
+            scalars = scalars.reshape(1, N)
         elif scalars.shape == (3, nlines) or scalars.shape == (nlines, 3):
             # one rgb color for each line, so broadcast it
             if scalars.shape == (3, nlines):
@@ -816,12 +825,20 @@ def prepare_lines(lines, scalars=None, do_connections=False):
                 colors.append(c)
             scalars = np.concatenate(colors, axis=1)
         else:
-            scalars = scalars.reshape(-1, np.sum(npts))
+            scalars = scalars.reshape(-1, N)
 
-        if scalars.shape[0] == 1:
+        if scalars.dtype.kind == 'S':
+            # translate hex colors (#ff00ff) into rgb values
+            scalars = np.char.lstrip(scalars, '#').astype('S6')
+            scalars = np.char.zfill(scalars, 6)
+            scalars = np.frombuffer(np.char.decode(scalars, 'hex'), dtype='u1')
+            scalars = scalars.reshape(-1, 3).T
+        elif scalars.shape[0] == 1:
+            # normal scalars
             scalars = scalars.reshape(-1)
         elif scalars.shape[0] == 3:
-            # The scalars encode rgb data, standardize the result
+            # The scalars encode rgb data, standardize the result to a
+            # 3xN ndarray of 1 byte unsigned ints (chars)
             if np.all(scalars >= 0) and np.all(scalars <= 1):
                 scalars = (255 * scalars).round().astype('u1')
             elif np.all(scalars >= 0) and np.all(scalars < 256):
@@ -834,6 +851,21 @@ def prepare_lines(lines, scalars=None, do_connections=False):
             raise ValueError("Scalars should either be a number, or set of "
                              "rgb values, shape is {0}".format(scalars.shape))
 
+    # broadcast / reshape additional arrays given in other
+    for key, arr in other.items():
+        if arr is None:
+            pass
+        elif arr.shape == (1, nlines) or arr.shape == (nlines, 1):
+            arr = arr.reshape(nlines, 1)
+            arr = [arr[i].repeat(ni) for i, ni in enumerate(npts)]
+            other[key] = np.concatenate(arr, axis=0).reshape(1, N)
+        else:
+            try:
+                other[key] = arr.reshape(-1, N)
+            except ValueError:
+                viscid.logger.warn("Unknown dimension, dropping array {0}"
+                                   "".format(key))
+
     if do_connections:
         connections = [None] * nlines
         for i, ni in enumerate(npts):
@@ -841,18 +873,18 @@ def prepare_lines(lines, scalars=None, do_connections=False):
             i0 = first_idx[i]
             connections[i] = np.vstack([np.arange(i0, i0 + ni - 1.5),
                                         np.arange(i0 + 1, i0 + ni - 0.5)]).T
-        connections = np.concatenate(connections, axis=0)
+        connections = np.concatenate(connections, axis=0).astype('i')
     else:
         connections = None
 
-    return lines, scalars, connections
+    return lines, scalars, connections, other
 
 def meshlab_convert(fname, fmt="dae", quiet=True):
     """Run meshlabserver to convert 3D mesh files
 
     Uses `MeshLab <http://meshlab.sourceforge.net/>`_, which is a great
     little program for playing with 3D meshes. The best part is that
-    OS X's Preview can open the COLLADA (*.dae) format. How cool is
+    OS X's Preview can open the COLLADA (`*.dae`) format. How cool is
     that?
 
     Args: