+++ CHANGES.md
@@ -12,6 +12,7 @@ Refactors:
   - blocks -> patches everywhere
   - block -> patch everywhere
   - vlab.multiplot argument nprocs -> nr_procs
+  - lots of names in `viscid.plot.mvi` and `viscid.calculator.topology`
 
 Bugfixes:
   - ibound topology detection (north/south accidently became dayside/nightside in version 0.90.0)